honey int->double 변경

+++ lib/models/user.dart
@@ -4,12 +4,13 @@ class User {
   String email;
   String provider;
   String thumbnail;
-  int honey = 0;
+  double honey = 0;
   int rank = -1;
   var works = [];
 
-  User(this.uid, this.displayName, this.email, this.provider, this.thumbnail, this.honey, this.rank, this.works);
-  
+  User(this.uid, this.displayName, this.email, this.provider, this.thumbnail,
+      this.honey, this.rank, this.works);
+
   @override
   String toString() {
     return "user.displayName: ${this.displayName}";
